# api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Don't change title and version, these are used in the code generation and should be consistent with the package.json version. It affects import paths and generated client code. You can update the description to reflect the current state of the API, but keep title and version as is for consistency.
  title: Api
  version: 0.1.0
  description: Rayo AI Personal Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /budget:
    get:
      operationId: getBudget
      tags: [budget]
      summary: Get all category budgets
      responses:
        "200":
          description: All category budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
        "404":
          description: No budget set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: upsertBudget
      tags: [budget]
      summary: Create or update budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertBudgetBody"
      responses:
        "200":
          description: Budget saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /budget/{category}:
    delete:
      operationId: deleteBudget
      tags: [budget]
      summary: Delete a budget category
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Budget not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        envelopeBased:
          type: boolean
      required:
        - id
        - name
        - email
        - envelopeBased
    UpdateMeBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        envelopeBased:
          type: boolean
...
    Budget:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        category:
          type: string
        monthlyLimit:
          type: number
        totalSpent:
          type: number
        remaining:
          type: number
        percentUsed:
          type: number
        rollover:
          type: boolean
      required:
        - id
        - userId
        - category
        - monthlyLimit
        - totalSpent
        - remaining
        - percentUsed
        - rollover
    UpsertBudgetBody:
      type: object
      properties:
        category:
          type: string
          minLength: 1
        monthlyLimit:
          type: number
          minimum: 0
        rollover:
          type: boolean
          default: false
        balance:
          type: number
          minimum: 0
      required:
        - category
        - monthlyLimit
        - rollover
...