# validation/generated/types/budget.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * Rayo AI Personal Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface Budget {
  id: number;
  userId: number;
  category: string;
  monthlyLimit: number;
  totalSpent: number;
  remaining: number;
  percentUsed: number;
  rollover: boolean;
}